animal_classifier/requirements.txt
absl-py==2.1.0
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioice==0.10.1
aiortc==1.9.0
aiosignal==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
aplr==10.7.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.5
attrs==24.2.0
audiomentations==0.42.0
audioread==3.0.1
av==12.3.0
babel==2.17.0
beautifulsoup4==4.12.3
bitsandbytes==0.46.1
bleach==6.2.0
blinker==1.8.2
blis==1.3.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
ChatterBot==1.2.7
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.1.0
cmudict==1.0.32
colorama==0.4.6
comm==0.2.2
comtypes==1.4.6
confection==0.1.5
contourpy==1.2.1
crcmod==1.7
cryptography==45.0.3
cycler==0.12.1
cymem==2.0.11
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_cytoscape==1.0.2
dataclasses-json==0.6.7
datasets==2.21.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.8.0
docarray==0.40.0
docstring_parser==0.16
editdistance==0.8.1
executing==2.0.1
fastapi==0.116.1
fastjsonschema==2.21.1
ffmpy==0.6.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.6.1
funasr==1.2.6
gast==0.6.0
gevent==24.10.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-crc32c==1.7.1
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.65.0
gradio==4.42.0
gradio_client==1.3.0
greenlet==3.1.1
groovy==0.1.2
grpcio==1.67.1
grpcio-status==1.62.3
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface==0.0.1
huggingface-hub==0.33.2
hydra-core==1.3.2
idna==3.7
ifaddr==0.2.0
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
intel-openmp==2021.4.0
interpret==0.6.5
interpret-core==0.6.5
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.7.0
jmespath==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaldiio==2.18.1
keras==3.9.2
keras-tuner==1.4.7
kiwisolver==1.4.5
kt-legacy==1.0.5
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.72
langchain-ollama==0.3.6
langchain-text-splitters==0.3.2
langcodes==3.5.0
langsmith==0.4.8
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llama-cloud==0.1.4
llama-index==0.11.22
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.22
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post4
llama-index-llms-groq==0.2.0
llama-index-llms-openai==0.2.16
llama-index-llms-openai-like==0.2.0
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.13
llvmlite==0.43.0
lxml==5.4.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
mathparse==0.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.18
mistune==3.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
modelscope==1.27.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.13.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
numpy-minmax==0.3.1
numpy-rms==0.4.2
ollama==0.5.1
omegaconf==2.3.0
openai==1.54.3
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.11
oss2==2.19.1
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.1
pooch==1.8.2
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.47
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.3
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
Pygments==2.18.0
pylibsrtp==0.12.0
pynndescent==0.5.13
pyOpenSSL==25.1.0
pyparsing==3.1.2
pypdf==4.3.1
PyPDF2==3.0.1
pyphen==0.17.2
pypiwin32==223
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-stretch==0.3.1
pytorch-wpe==0.0.1
pyttsx3==2.91
pytz==2024.1
pywin32==306
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.1.1
referencing==0.35.1
regex==2024.11.6
reportlab==4.4.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.1
rsa==4.9
ruff==0.12.1
safetensors==0.4.5
SALib==1.5.1
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.2.1
sentencepiece==0.2.0
setuptools==80.9.0
shap==0.46.0
shellingham==1.5.4
shtab==1.7.2
simsimd==6.2.1
six==1.16.0
slicer==0.0.8
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.35
srsly==2.5.1
srt==3.5.3
stack-data==0.6.3
starlette==0.47.3
streamlit==1.39.0
streamlit-webrtc==0.47.7
stringzilla==3.12.3
striprtf==0.0.26
sympy==1.14.0
tbb==2021.13.1
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorflow-intel==2.16.2
termcolor==2.4.0
terminado==0.18.1
textblob==0.19.0
textstat==0.7.5
thinc==8.3.6
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
torch==2.4.1
torch-complex==0.4.4
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
trl==0.9.6
typeguard==4.4.4
typer==0.12.5
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.14.1
tyro==0.9.26
tzdata==2024.1
ultralytics==8.3.177
ultralytics-thop==2.0.15
umap-learn==0.5.7
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.35.0
vosk==0.3.45
wasabi==1.1.3
watchdog==5.0.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.4
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.16.0
xarray==2024.10.0
xgboost==3.0.0
xxhash==3.5.0
yarl==1.17.1
yt-dlp==2025.6.25
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.1
zstandard==0.23.0
